Lambda/requirements.txt
spotipy == 2.20
boto3 == 1.24
botocore == 1.27
pypi-json == 0.3
redis == 4.3
index == 0.2
async-timeout==3.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
idna==2.10
importlib-metadata==4.6.4
jmespath==0.10.0
platformdirs==2.4.1
py==1.10.0
pyparsing==2.4.7
s3fs==2021.7.0
s3transfer==0.5.0
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.6
wrapt==1.12.1
packaging == 21.3
zipp== 3.8.1
PyGithub== 1.55
pyjwt, 
pycparser, 
cffi, 
pynacl,
pandas,
numpy,
pytz